src/main/java/com/leetcode/algorithm/easy/PathSum.java
package com.leetcode.algorithm.easy;

import com.leetcode.algorithm.common.struct.TreeNode;
import lombok.extern.slf4j.Slf4j;

/**
 * 112. Path Sum
 * https://leetcode.com/problems/path-sum/
 *
 * Definition for a binary tree node.
 * public class TreeNode {
 *     int val;
 *     TreeNode left;
 *     TreeNode right;
 *     TreeNode() {}
 *     TreeNode(int val) { this.val = val; }
 *     TreeNode(int val, TreeNode left, TreeNode right) {
 *         this.val = val;
 *         this.left = left;
 *         this.right = right;
 *     }
 * }
 */
@Slf4j
public class PathSum {
    public boolean hasPathSum(TreeNode root, int sum) {
        if (root.val == sum) {
            return true;
        }


        return false;
    }
}
